refactor(frontend): remove TRPCExample component and implement useConnectWallet hook for wallet connection
