+++ packages/frontend/src/features/home/pages/home-page.tsx
@@ -1,11 +1,9 @@
 import Header from '../components/header';
-import { TRPCExample } from '@/components/trpc-example';
 
 export default function HomePage() {
   return (
     <main>
       <Header />
-      <TRPCExample />
     </main>
   );
 }
